Fixed bug with iDegeneratePairScore() and iDegenerateSingletScore() functions involving the new scaled int log odds scores in alphabet.c.
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/infernal/trunk@1779 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/funcs.h
@@ -19,8 +19,8 @@ extern float  DegeneratePairScore(float *esc, char syml, char symr);
 extern float  DegenerateSingletScore(float *esc, char sym);
 extern char  *DigitizeSequence(char *seq, int L);
 extern char **DigitizeAlignment(char **aseq, int nseq, int alen);
-extern int    iDegeneratePairScore(float *esc, char syml, char symr);
-extern int    iDegenerateSingletScore(float *esc, char sym);
+extern int    iDegeneratePairScore(int *esc, char syml, char symr);
+extern int    iDegenerateSingletScore(int *esc, char sym);
 
 /* from bandcyk.c
  */
